컴포넌트 분리 밒 다이얼로그 분리
